Continuing with BoxFactoryUpdated
I have added the new directory system. There is a class that runs menus
based off of the directory location. I would like to move all the
directory calls from the menus to the directory class. This would make
the Menus class only for menus.

I also would like to add an image object with child objects such as a
gif in order to streamline the display code.